Code review feedback 4.
